refactor(metal-vm): Move the power operations to their own file
